Refactor: Organize Rocket component file structure
`Rocket`コンポーネントのファイル構造を見直し、`/Rocket`ディレクトリに移動しました。この変更により、関連ファイルが一箇所に集約され、コードの可読性と管理性が向上しました。また、インポートパスを調整し、他の部分での参照を適切に修正しました。
