+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import Footer from '@/components/Footer'
 import Copyright from '@/components/Copyright'
 import ModuleSelector from '@/components/ModuleSelector'
 import Distance from '@/components/Distance'
-import Rocket from '@/components/Rocket'
+import Rocket from '@/components/Rocket/Rocket'
 import ResetButton from '@/components/ResetButton'
 import Results from '@/components/Results/Results'
 import Header from '@/app/Header'
